+++ HaruhiHeiretsuLib/Strings/ScriptFile.cs
@@ -31,7 +31,7 @@ namespace HaruhiHeiretsuLib.Strings
         {
         }
 
-        public ScriptFile(int parent, int child, byte[] data, int mcbId = 0)
+        public ScriptFile(int parent, int child, byte[] data, short mcbId = 0)
         {
             Location = (parent, child);
             McbId = mcbId;
